pkp/pkp-lib#2163 Move focus/blur handlers for keyboard nav to ListPanelItem
