+++ js/controllers/list/ListPanelItem.vue
@@ -1,5 +1,5 @@
 <template>
-	<li class="pkpListPanelItem" :class="{'--isFeatured': isFeatured}">
+	<li class="pkpListPanelItem" :class="{'--has-focus': isFocused}">
 		<div class="pkpListPanelItem--submission__item"></div>
 	</li>
 </template>
@@ -11,9 +11,24 @@ export default {
 		return {
 			// ListPanelItem components must have an id for ordering to work
 			id: null,
+			isFocused: false,
 		}
 	},
 	methods: {
+		/**
+		 * Update the isFocused property
+		 */
+		focusItem: function() {
+			this.isFocused = true;
+		},
+
+		/**
+		 * Update the isFocused property
+		 */
+		blurItem: function() {
+			this.isFocused = false;
+		},
+
 		/**
 		 * Pass an itemOrderUp event up to the list panel
 		 *
